Radical restructuring to create package
