+++ .gitignore
@@ -1,104 +1,12 @@
-# Byte-compiled / optimized / DLL files
+# Directories
 __pycache__/
-*.py[cod]
 
-# C and Fortran extensions
-*.so
-*.dylib
-*.dll
-*.slo
-*.lo
-*.o
-*.obj
-*.mod
-*.smod
-*.lai
-*.la
-*.a
-*.lib
-*.exe
-*.out
-*.app
 
-# Distribution / packaging
-.Python
-env/
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
+# setuptools_scm
+resmda/version.py
+
+# Build related
 .eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-*.egg-info/
-.installed.cfg
-*.egg
-
-# PyInstaller
-#  Usually these files are written by a python script from a template
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
-*.manifest
-*.spec
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*,cover
-
-# Translations
-*.mo
-*.pot
-
-# Django stuff:
-*.log
-
-# Sphinx documentation
-docs/_build/
-
-# PyBuilder
-target/
-
-# DotEnv configuration
-.env
-
-# Database
-*.db
-*.rdb
-
-# Pycharm
-.idea
-
-# TeX/LaTeX
-*.aux
-*.bbl
-*.blg
-*.synctex.gz
-*.xwm
-
-# IPython NB Checkpoints
-.ipynb_checkpoints/
-
-# exclude compiled binaries
-bin/
-
-# exclude data from source control by default
-/data/external
-/data/interim
-/data/processed
-/data/raw
-
-# exclude external models from source control
-/src/external/
+build/
+dist/
+resmda.egg-info/
