--- AUTHORS.md
@@ -1,12 +1,0 @@
-Credits
-=======
-
-Project Lead
-----------------
-
-* tuda-geo  <g.serraoseabra@tudelft.nl>
-
-Project Contributors
-------------
-
-None.
