+++ README.md
@@ -1,29 +1,14 @@
-da-em-ccs
-==============================
+# `resmda` - a simple reservoir modeller with ESMDA
 
-Data assimilation with EM for CCS
 
-Project Organization
---------------------
+The package `resmda` contains a simple, 2D reservoir simulator and a
+straight-forward implementation of the basic *Ensemble smoother with multiple
+data assimilation* (ESMDA) algorithm as presented by Emerick and Reynolds,
+2013.
 
-    .
-    ├── AUTHORS.md
-    ├── LICENSE
-    ├── README.md
-    ├── bin
-    ├── config
-    ├── data
-    │   ├── external
-    │   ├── interim
-    │   ├── processed
-    │   └── raw
-    ├── docs
-    ├── notebooks
-    ├── reports
-    │   └── figures
-    └── src
-        ├── data
-        ├── external
-        ├── models
-        ├── tools
-        └── visualization
+
+## Installation
+
+```
+pip install git+https://github.com/tuda-geo/resmda.git
+```
