+++ setup.cfg
@@ -1,0 +1,2 @@
+[flake8]
+per-file-ignores = __init__.py: F401
